feat(editor): read the graph through reflection instead of a snapshot
Group 7 of redesign-graph-model. The editor populates its widgets from a
borrow of `&Graph` plus the type registry, held only for the duration of
one call -- no Arc, no mutex, no copy, and no hand-written view layer
between the graph and the widgets (design D11).

Reflection is the schema rather than something shadowed by one. A field
row carries the field's own `&'static TypeInfo`; the control is chosen
from it and the committed text is parsed back against the same one, so
the two cannot disagree. Sockets come from the node kind's declared
inlets and outlets, which is what gives an unconnected inlet a socket,
and the palette is a read of the registered kinds rather than a
`ComponentDocRegistry`.

Sockets and edges are addressed by `(NodeId, field path)`, resolved by
key everywhere, so re-sorting a node's sockets moves the dot and takes
the edge with it. Drag-to-connect decides legality against the same type
information `Graph::connect` compares and never sends a connection it
knows would be refused; a drop onto an occupied single-connection inlet
is sent, because that replaces rather than fails.

Variadic inlets are reordered from the inspector, which needed a row type
rather than a new interaction model -- the canvas fan has no specified
visual design, and inventing one is not this change's business.

The old snapshot path stays until sway-app switches over; group 9 deletes
`snapshot.rs`, so task 7.1 stays open until then.

+++ crates/sway-editor/src/viewport.rs
@@ -175,9 +175,7 @@ impl Widget for Viewport {
                 gesture: PointerGesture::Pinch(delta),
                 ..
             }) => {
-                self.send(ViewportInput::Pinch {
-                    delta: *delta as f32,
-                });
+                self.send(ViewportInput::Pinch { delta: *delta });
                 ctx.set_handled();
             }
             _ => {}
